pulling master from exec

+++ inc/types.hpp
@@ -4,13 +4,13 @@
 #include <cstdint>
 
 // Tuple types:
-using tuple_key_t     = int32_t;
-using tuple_payload_t = int32_t;
+using tuple_key_t     = int64_t;
+using tuple_payload_t = int64_t;
 
 // Map types:
-using bucket_key_t    = int32_t;
+using bucket_key_t    = int64_t;
 
-using chain_key_t     = int32_t;
+using chain_key_t     = int64_t;
 
 // PsumTable types:
 using radix_t         = uint8_t;
